Type-correctness

+++ device.go
@@ -48,7 +48,7 @@ func FetchDevices(c *Client) ([]*Device, error) {
 	}
 
 	ds := make([]*Device, deviceCount)
-	for i := 0; i < deviceCount; i++ {
+	for i := uint32(0); i < deviceCount; i++ {
 		ds[i], err = fetchDevice(c, i)
 		if err != nil {
 			return []*Device{}, fmt.Errorf("Couldn't fetch Device %d: %w", i, err)
@@ -63,7 +63,7 @@ func FetchDevices(c *Client) ([]*Device, error) {
 * Devices' length is fetched by a separate command.
 * Then each device by a command.
 * Within a device, zones etc aren't API commands, they're packed into the binary blob, with thier length preceeding them */
-func fetchDeviceCount(c *Client) (int, error) {
+func fetchDeviceCount(c *Client) (uint32, error) {
 	if err := c.sendCommand(0, cmdGetDevCnt, []byte{}); err != nil {
 		return 0, fmt.Errorf("Couldn't fetch Device count: %w", err)
 	}
@@ -73,12 +73,12 @@ func fetchDeviceCount(c *Client) (int, error) {
 		return 0, fmt.Errorf("Couldn't fetch Device count: %w", err)
 	}
 
-	deviceCount := int(binary.LittleEndian.Uint32(body))
+	deviceCount := binary.LittleEndian.Uint32(body)
 
 	return deviceCount, nil
 
 }
-func fetchDevice(c *Client, i int) (*Device, error) {
+func fetchDevice(c *Client, i uint32) (*Device, error) {
 	if err := c.sendCommand(uint32(i), cmdGetDevData, []byte{}); err != nil {
 		return &Device{}, fmt.Errorf("Couldn't fetch Device %d: %w", i, err)
 	}
